refactored otel envar config processor init into dedicated functions
